LayoutTests/media/modern-media-controls/macos-inline-media-controls/macos-inline-media-controls-start-button-style-expected.txt

Testing the MacOSInlineMediaControls computed styles for the start button.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getComputedStyle(mediaControls.playPauseButton.element.firstElementChild.firstElementChild).webkitBackdropFilter is "saturate(1.8) blur(17.5px)"
PASS successfullyParsed is true

TEST COMPLETE
